# lib/api-client-react/src/generated/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface ProfileField {
  label: string;
  value: string;
}

export interface ExtractedProfile {
  displayName?: string;
  currentDegree?: string;
  institution?: string;
  yearOfStudy?: string;
  skills?: string;
  city?: string;
  preferredIndustries?: string;
  careerGoals?: string;
  portfolioUrl?: string;
  profileFields?: ProfileField[];
}

export interface ProfileChatInput {
  messages: ChatMessage[];
  existingProfile?: ExtractedProfile;
  /** Extracted text from an uploaded CV/Resume — the AI treats this as pre-known information and skips asking about things it can already read */
  cvContent?: string;
}

export interface ParseProfileFromCvRequest {
  /** Raw extracted text from a CV or Resume document */
  cvContent: string;
}

export interface ProfileChatResult {
  reply: string;
  isComplete: boolean;
  profileData?: ExtractedProfile;
}

export interface DiscoverCompaniesInput {
  latitude: number;
  longitude: number;
  degree: string;
  institution?: string;
  yearOfStudy?: string;
  skills?: string;
  city?: string;
  preferredIndustries?: string;
  goals?: string;
  /** Extracted text from the student's uploaded documents (CV, certificates, etc.) to enrich matching */
  documentsContext?: string;
}

export interface CompanyDiscovery {
  name: string;
  description: string;
  fitScore: string;
  /** @nullable */
  website?: string | null;
}

export interface DraftLetterInput {
  companyName: string;
  role: string;
  degree: string;
  goals: string;
  institution?: string;
  yearOfStudy?: string;
  skills?: string;
  portfolioUrl?: string;
  userDraft?: string;
  /** Type of letter: 'attachment' for industrial attachment, 'internship', 'graduate' for graduate programme, 'general' for general application */
  letterType?: string;
  studentName?: string;
  studentPhone?: string;
  studentEmail?: string;
  studentCity?: string;
  /** Extracted text from the student's CV/Resume to personalise the letter with real experience and achievements */
  cvContent?: string;
}

export interface DraftLetterResult {
  letter: string;
}

export interface ResearchCompanyInput {
  companyName: string;
}

export interface ResearchCompanyResult {
  summary: string;
}

export interface StarFeedbackInput {
  question: string;
  situation: string;
  task: string;
  action: string;
  result: string;
}

export interface StarFeedbackResult {
  feedback: string;
}

export interface InterviewQuestionsInput {
  companyName: string;
  role: string;
  degree: string;
  goals: string;
  institution?: string;
  yearOfStudy?: string;
  skills?: string;
  researchSummary?: string;
  /** Extracted text from the student's CV/Resume to tailor questions to their actual experience */
  cvContent?: string;
}

export interface NetworkingEventsInput {
  city: string;
  degree?: string;
  preferredIndustries?: string;
  goals?: string;
}

export interface NetworkingEvent {
  id: string;
  title: string;
  eventType: string;
  organizer: string;
  dateLabel: string;
  dateIso?: string;
  location: string;
  description?: string;
  url?: string;
  source?: string;
  tags?: string[];
  isOnline?: boolean;
}

export interface InterviewQuestionsResult {
  personal: string[];
  company: string[];
  experience: string[];
}

export interface InterviewVerdictInput {
  companyName: string;
  role: string;
  degree: string;
  goals: string;
  institution?: string;
  yearOfStudy?: string;
  skills?: string;
  city?: string;
  questions: string[];
  answers: string[];
  researchSummary?: string;
  /** Extracted text from the student's CV/Resume to evaluate answers against claimed experience */
  cvContent?: string;
}

export interface AnswerFeedback {
  question: string;
  answer: string;
  feedback: string;
  score: number;
}

export interface InterviewVerdictResult {
  /** accepted | shortlisted | rejected */
  verdict: string;
  overallScore: number;
  overallFeedback: string;
  strengths: string[];
  areasToImprove: string[];
  answerFeedback: AnswerFeedback[];
  recommendation: string;
}

export interface ExtractContentRequest {
  /** Normalized object path (e.g. /objects/uploads/uuid) */
  objectPath: string;
  /** MIME type of the file */
  contentType: string;
  /** Document category (e.g. CV / Resume, Certificate) */
  category: string;
}

export interface ExtractContentResult {
  /** Career-relevant text extracted from the document */
  extractedText: string;
}

export interface UploadUrlRequest {
  /**
   * Original file name.
   * @minLength 1
   */
  name: string;
  /**
   * File size in bytes.
   * @minimum 1
   */
  size: number;
  /**
   * MIME type of the file.
   * @minLength 1
   */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. /objects/uploads/uuid). */
  objectPath: string;
}

export interface ErrorEnvelope {
  error: string;
}