# lib/api-zod/src/generated/generated/types/extractContentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractContentRequest {
  /** Normalized object path (e.g. /objects/uploads/uuid) */
  objectPath: string;
  /** MIME type of the file */
  contentType: string;
  /** Document category (e.g. CV / Resume, Certificate) */
  category: string;
}