# lib/api-zod/src/generated/generated/types/uploadUrlRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlRequest {
  /**
   * Original file name.
   * @minLength 1
   */
  name: string;
  /**
   * File size in bytes.
   * @minimum 1
   */
  size: number;
  /**
   * MIME type of the file.
   * @minLength 1
   */
  contentType: string;
}